FIX: db_merge.c (db_merge): 'in-place' error
When using the option '--in-place', all the databases are merged into
the first one of the list. In order to avoid that the same exon is
stored twice, we index the values in a hash. An error arise about the
'ense id', because the values already stored in the database to merge
to are not taken into acount.

Try to fix it indexing the pre-recorded values before starting to merge
the databases.